requirements.txt
async-generator==1.10
attrs==22.1.0
beautifulsoup4==4.11.1
bidict==0.22.0
certifi==2022.9.14
cffi==1.15.1
click==8.1.3
cryptography==38.0.1
distlib==0.3.6
dnspython==2.2.1
eventlet==0.33.2
exceptiongroup==1.0.0rc9
filelock==3.8.0
Flask==2.2.2
Flask-SocketIO==5.3.2
Flask-SQLAlchemy==3.0.2
greenlet==1.1.3.post0
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jwt==1.3.1
MarkupSafe==2.1.1
mysql-connector-python==8.0.31
numpy==1.23.2
outcome==1.2.0
pipenv==2022.11.25
platformdirs==2.5.4
protobuf==3.20.1
pycparser==2.21
PyJWT==2.6.0
pymongo==4.2.0
PyMySQL==1.0.2
PySocks==1.7.1
python-dotenv==0.21.0
python-engineio==4.3.4
python-socketio==5.7.2
selenium==4.5.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.42
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.12
virtualenv==20.16.7
virtualenv-clone==0.5.7
websocket-client==1.4.2
Werkzeug==2.2.2
wsproto==1.2.0
